chore(scripts): extract shared import-rewriter to scripts/lib
The two postbuild scripts had ~40 lines of identical regex-walker code for
adding .js extensions to bare relative imports. Lifted into
`scripts/lib/rewrite-relative-imports.mjs` so both packages share one
implementation. Per-package scripts now hold only the bespoke parts
(relocate / .d.cts duplication for nimbus; dual-package markers for icons).

Also renamed `packages/nimbus-icons/scripts/write-dual-package-markers.mjs`
→ `postbuild.mjs` since it does more than markers now, and updated the
package.json script name `build:markers` → `build:postbuild` for symmetry
with nimbus's `build:postbuild-types`.

Considered `tsc-esm-fix` as a drop-in. Works for nimbus-icons (standard
tsc 1:1 .js+.d.ts shape) but can't handle nimbus — vite-plugin-dts emits
.d.ts files in directory trees without companion .js siblings, and
tsc-esm-fix's resolver requires those siblings to confirm an import is
valid. Keeping our own resolver lets one approach cover both packages.

Verified: pnpm check:package-shape (strict mode, all green), pnpm lint,
pnpm test:unit (1180 tests).
