+++ packages/nimbus-icons/package.json
@@ -37,11 +37,11 @@
     "typescript": "catalog:tooling"
   },
   "scripts": {
-    "build": "pnpm run build:icons && pnpm run build:esm && pnpm run build:cjs && pnpm run build:markers",
+    "build": "pnpm run build:icons && pnpm run build:esm && pnpm run build:cjs && pnpm run build:postbuild",
     "build:icons": "pnpm svgr --filename-case kebab --typescript --jsx-runtime automatic --out-dir src/material-icons node_modules/@material-design-icons/svg/outlined",
     "build:esm": "tsc --project tsconfig.esm.json",
     "build:cjs": "tsc --project tsconfig.cjs.json",
-    "build:markers": "node scripts/write-dual-package-markers.mjs",
+    "build:postbuild": "node scripts/postbuild.mjs",
     "dev": "pnpm run build:esm && pnpm run build:cjs"
   }
 }
